SpuriousZero is now MangosVBZero.
I changed all the references to Spurious to MangosVB, also changed the license at the top of modules to the getMangos version.  Added the COPYING and WARNING files to the Documentation directory.

Removed the SpuriousZero. from the front of the Common dll, and the servers exe files.

Changed the build configuration to build using Debug x86 instead of Debug AnyCPU, so we would not have the invalid format problems with the libeay32.dll on 64 bit machines.

Upgraded all the Tools projects to VS2010, the Spells Browser still needs codes and dbc stuff updated from 1.11 to 1.12 versions.