Added CommandCharacter element to WorldServer.ini file to tell the core what character to use for system commands in the chat.
 If you put a . in the CommandCharacter element, then you would do system commands like .info to get info on a targeted npc, if you put an ! in the CommandCharacter you would do the command like !info.

Fixed the field names used in the query for the gotostart command.  Added default ai to the npcflood command.

Changed the Help command to list each command on a separate line, the way it was, would not fit in the chat area, and you could not see the list after it showed.

Fixed a couple of typos in comments.