Applied Standard Mangos DB Startup checks
The standard DB Version format is Rel xx.yy.zzz

xx  = The Main Release Number
yy  = The Database structure Revision, this changes whenever the
      DB format changes
zzz = The Content Revision, typically changes each db update

A mismatch with either xx or yy is regarded as a Fatal error and
should prevent the server from starting.

A mismatch with zzz is only regarded as a warning. The server will
still start up.